refactor: adjust imports and exports for features
