+++ src/features/channels/index.tsx
@@ -1,0 +1,2 @@
+export * from "./components";
+export * from "./channelsSlice";
